Fixes syntax error for 1.8.7
The only use of Aruba::Config::Hooks#execute was in `api.rb` and
that was passing a single value for the second argument.

So, I removed the splat in the arg and the call to
`instance_exec`.  Since none of this is documented, and no tests
broke, I'm going to assume this is OK, at least to get
a working version for 1.8.7.