Demonstration of using polymorphism for matchers
@maxmeyer here's what I was talking about in #238. Just a suggestion -
WDYT?

Note that I had to pass in a reference to `aruba` so that I could still
get to the `absolute_path` helper method. Maybe we can put this in a
module that can be mixed into the helpers if we like this way of doing
things...?

@myronmarston I'd appreciate your feedback about whether there's a
neater way to do this if you have time - particularly the namespacing of
the factory method - I used RSpec::Matchers just because I assumed
that's where the DSL puts them.