Fix missing variable in exception handler
Previously, "cmd" was passed to a different method.

The reason this never came up earlier is because on Unix, if a file is
not in the PATH, a Aruba::LaunchError is raised much earlier.

One way this could've been triggered is: if the program file existed,
but wasn't accessible (e.g. EACCES error). That way it wouldn't fail the
`which` test, but would fail once it was attempted to be executed.