Fix JRuby before :command helper hook
The before :command hooks are run after the command's environment has
already been set up, so instead of modifying Aruba's environment, of the
actual environment, we need to modify the command's environment. This
change adds a scenario for this feature, and updates the implementation
to make it work.

We also no longer require the helper by default, so requiring by the
user will have the desired effect.