Find relative commands from the current directory
In an Arabu spec or scenario, files are created relative to the current
Aruba directory. When creating an executable in that way, it should be
possible to run that executable using the same relative path.

This change means that executables in the project under test can no
longer be find using the relative path from the project directory. They
should be added to the path. This is done automatically by Bundler.