Allow newer versions of Cucumber
This allows us to test / release new versions of Cucumber without
first having to come here to make a new release of Arubua.

I think the only reason to cap the max version would be if we knew there
was a backwards-compatibility problem with a release of Cucumber, but
there isn't, so let's remove it?