Use plain irb for debugging
Development of pry has stalled and it is suggested that debug in
combination with irb is the future. For now, we cannot use debug yet
since only a very old version runs on Ruby 2.5.

For now, this change removes the pry, pry-doc, byebug and pry-byebug
dependencies. These can be added back via Gemfile.local.

Once support for Ruby 2.5 has been dropped, the new debug gem can be
added as a permanent dependency to provide extended debugging features
by default.