Make scenario for stopping on output faster more robust
By changing the printed text, the first iteration will already print
text that is not part of the echoed command. Therefore, we can trigger
termination after the first iteration instead of the second.

To avoid slower termination failing the scenario, allow for Test 1 to be
output, and just require Test 2 not to appear.