Restore coverage collection for cucumber features
Note: `simplecov_setup` has been moved to it's own directory because of the way that it needs to be specified when setting the `RUBYOPTS` environment variable. Just specifying the path to the full file path via `-r` does not work. When doing that an error is generated that the file cannot be found. Instead, the directory containing the file is added to the load path via `-I`, and then the file is required with a simple `-rsimplecov_setup`. Because this modifies the load path, having a file named `aruba.rb` in the same directory as `simplecov_setup.rb` caused problems for files that attempted to `require 'aruba'`.