Suppress console output generated by Simplecov HTML formatter
The Simplecov HTML formatter, which is the default, writes a message to the console after it has stored coverage information the `coverage` directory. There is no built in way to suppress this message. See simplecov-ruby/simplecov-html#116 and https://github.com/simplecov-ruby/simplecov-html/blob/v0.12.3/lib/simplecov-html.rb#L23..L32.

The console output generated by this formatter causes failures for tests that are expecting exact command output.

The approach employed here is to monkey patch the method so that it still generates the same message, but instead of writing to the console, it writes to a file in the `coverage` directory named `summary.txt`.