Use plain irb in debugging scenario
In Ruby 3.3, requiring readline does not load the real readline library
but falls back to reline by default. This creates odd output when the
terminal is not a real TTY and a prompt is used when getting input.
Because pry uses a prompt, this makes the interactive debugging scenario
fail.

This change works around this issue by switching to irb. Irb will not
use a prompt, possibly because the output is not a TTY.

See https://github.com/cucumber/aruba/issues/910. Also, see
https://github.com/ruby/reline/issues/616 for the resulting reline
issue.