Unlink Tempfiles after use in SpawnProcess
On Windows, the finalizer for Tempfile will raise an error if
Errno::ENOACCES is raised when removing the file. However, if this
happens when #unlink is called, the error is caught.

This change calls #unlink on the Tempfiles that are used to store the
output of spawned processes. This means the finalizer will not need to
unlink the file, so any Errno::ENOACCES errors during unlinking will be
caught.

The raised errors did not make any specs fail, but they did appear in
the rspec output.