Fix simplecov setup for cucumber features
The way Aruba's environment manipulation currently works, when running
code or child processes in a nested environment, the priority of how the
environment variables are set is like this:

  1. Variables set with Aruba's environment manipulation methods
  2. Variables set in ENV when Aruba's environment manipulation object
     (aruba.environment) was initialized
  3. Variables set in ENV after aruba.environment was initialized

I'm not convinced the order of 2. and 3. is the right order, but due to
this order, setting RUBYOPT directly on ENV had no effect because there
was already a variable RUBYOPT set at the point of initializion of
aruba.environment.

By using the environment manipulation methods, the values set here will
take first priority regardless of whether the variables were set
earlier.